se añadi+o initial screen
