+++ src/screens/InitialScreen/InitialScreen.jsx
@@ -1,18 +1,8 @@
 import React from 'react'
 import { StatusBar } from 'expo-status-bar'
-import {
-  StyleSheet, Text, View, Button,
-} from 'react-native'
+import { Text, View, Button } from 'react-native'
 import { useNavigation } from '@react-navigation/native'
-
-const styles = StyleSheet.create({
-  container: {
-    alignItems: 'center',
-    backgroundColor: '#fff',
-    flex: 1,
-    justifyContent: 'center',
-  },
-})
+import styles from './styles'
 
 const App = () => {
   const navigation = useNavigation()
