+++ src/screens/InitialScreen/index.js
@@ -1,0 +1,3 @@
+import InitialScreen from './InitialScreen'
+
+export default InitialScreen
