+++ src/screens/InitialScreen/styles.js
@@ -1,0 +1,10 @@
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    backgroundColor: '#fff',
+    flex: 1,
+    justifyContent: 'center',
+  },
+})
+
+export default styles
